[LoRA] Centralize LoRA tests
